+++ tests/databases/AWSService.test.js
@@ -1,4 +1,4 @@
-import { AWSService } from "../../databases/awsService";
+import { AWSService } from "../../repository/services/awsService";
 import sinon from 'sinon';
 import AWS from 'aws-sdk';
 import { expect } from 'chai';
